+++ app/src/components/settings/panels/VoicePanel.tsx
@@ -20,9 +20,9 @@ const VoicePanel = () => {
   const [settings, setSettings] = useState<VoiceServerSettings | null>(null);
   const [savedSettings, setSavedSettings] = useState<VoiceServerSettings | null>(null);
   const [serverStatus, setServerStatus] = useState<VoiceServerStatus | null>(null);
-  const [voiceStatus, setVoiceStatus] = useState<VoiceStatus | null>(null);
+  const [, setVoiceStatus] = useState<VoiceStatus | null>(null);
   const [sttReady, setSttReady] = useState(false);
-  const [isLoading, setIsLoading] = useState(true);
+  const [, setIsLoading] = useState(true);
   const [isSaving, setIsSaving] = useState(false);
   const [isStarting, setIsStarting] = useState(false);
   const [isStopping, setIsStopping] = useState(false);
@@ -53,9 +53,6 @@ const VoicePanel = () => {
         openhumanVoiceStatus(),
         openhumanLocalAiAssetsStatus(),
       ]);
-      // Only overwrite local settings if there are no unsaved edits,
-      // or if explicitly forced (e.g. after save or initial load).
-      // This prevents the 2s polling timer from clobbering user input.
       const currentSettings = settingsRef.current;
       const currentSavedSettings = savedSettingsRef.current;
       if (
@@ -191,60 +188,12 @@ const VoicePanel = () => {
       />
 
       <div className="p-4 space-y-4">
-        <section className="space-y-3">
-          <div className="bg-stone-50 rounded-lg border border-stone-200 p-4 space-y-3">
-            <div className="flex items-center justify-between">
-              <div>
-                <h3 className="text-sm font-semibold text-stone-900">Runtime</h3>
-                <p className="text-xs text-stone-500 mt-1">
-                  Hold the hotkey to dictate and insert text into the active field.
-                </p>
-              </div>
-              <button
-                type="button"
-                onClick={() => void loadData()}
-                className="text-xs text-primary-600 hover:text-primary-700">
-                Refresh
-              </button>
-            </div>
-
-            <div className="grid grid-cols-2 gap-3 text-sm">
-              <div className="rounded-md border border-stone-200 bg-white p-3">
-                <div className="text-[10px] uppercase tracking-wide text-stone-500">Server</div>
-                <div className="mt-1 font-medium text-stone-900">
-                  {serverStatus ? serverStatus.state : isLoading ? 'Loading…' : 'Unavailable'}
-                </div>
-              </div>
-              <div className="rounded-md border border-stone-200 bg-white p-3">
-                <div className="text-[10px] uppercase tracking-wide text-stone-500">STT</div>
-                <div className="mt-1 font-medium text-stone-900">
-                  {voiceStatus?.stt_available ? 'Ready' : 'Not ready'}
-                </div>
-              </div>
-            </div>
-
-            {serverStatus && (
-              <div className="grid grid-cols-1 sm:grid-cols-3 gap-3 text-xs text-stone-600">
-                <div>Hotkey: {serverStatus.hotkey || 'n/a'}</div>
-                <div>Mode: {serverStatus.activation_mode}</div>
-                <div>Transcriptions: {serverStatus.transcription_count}</div>
-              </div>
-            )}
-
-            {serverStatus?.last_error && (
-              <div className="rounded-md border border-red-200 bg-red-50 p-3 text-xs text-red-600">
-                {serverStatus.last_error}
-              </div>
-            )}
-          </div>
-        </section>
-
         <section className={`space-y-3 ${disabled ? 'opacity-60' : ''}`}>
           <div className="bg-stone-50 rounded-lg border border-stone-200 p-4 space-y-4">
             <div>
-              <h3 className="text-sm font-semibold text-stone-900">Voice Server Settings</h3>
+              <h3 className="text-sm font-semibold text-stone-900">Voice Settings</h3>
               <p className="text-xs text-stone-500 mt-1">
-                Configure startup behavior, hotkey handling, and transcription cleanup.
+                Hold the hotkey to dictate and insert text into the active field.
               </p>
             </div>
 
@@ -286,53 +235,14 @@ const VoicePanel = () => {
                   </label>
                 </div>
 
-                <div className="grid grid-cols-1 sm:grid-cols-2 gap-4">
-                  <label className="flex items-center gap-2 text-sm text-stone-700">
-                    <input
-                      type="checkbox"
-                      checked={settings.auto_start}
-                      onChange={e => updateSetting('auto_start', e.target.checked)}
-                      className="h-4 w-4 rounded border-stone-300 text-primary-600 focus:ring-primary-500"
-                    />
-                    Start voice server automatically with the core
-                  </label>
-
-                  <label className="block space-y-1">
-                    <span className="text-xs font-medium text-stone-600">
-                      Minimum Recording Seconds
-                    </span>
-                    <input
-                      type="number"
-                      min="0"
-                      step="0.1"
-                      value={settings.min_duration_secs}
-                      onChange={e =>
-                        updateSetting('min_duration_secs', Number(e.target.value) || 0)
-                      }
-                      className="w-full rounded-md border border-stone-200 bg-white px-3 py-2 text-sm text-stone-900 focus:outline-none focus:ring-1 focus:ring-primary-400"
-                    />
-                  </label>
-                </div>
-
-                <label className="block space-y-1">
-                  <span className="text-xs font-medium text-stone-600">
-                    Silence Threshold (RMS)
-                  </span>
-                  <p className="text-[11px] text-stone-400">
-                    Recordings with energy below this are treated as silence and skipped. Lower =
-                    more sensitive.
-                  </p>
+                <label className="flex items-center gap-2 text-sm text-stone-700">
                   <input
-                    type="number"
-                    min="0"
-                    max="1"
-                    step="0.001"
-                    value={settings.silence_threshold}
-                    onChange={e =>
-                      updateSetting('silence_threshold', Number(e.target.value) || 0.002)
-                    }
-                    className="w-full rounded-md border border-stone-200 bg-white px-3 py-2 text-sm text-stone-900 focus:outline-none focus:ring-1 focus:ring-primary-400"
+                    type="checkbox"
+                    checked={settings.auto_start}
+                    onChange={e => updateSetting('auto_start', e.target.checked)}
+                    className="h-4 w-4 rounded border-stone-300 text-primary-600 focus:ring-primary-500"
                   />
+                  Start voice server automatically with the core
                 </label>
 
                 <div className="space-y-2">
@@ -452,6 +362,16 @@ const VoicePanel = () => {
             </div>
           </div>
         </section>
+
+        <button
+          type="button"
+          onClick={() => navigateToSettings('voice-debug')}
+          className="flex items-center gap-1.5 text-xs text-stone-400 hover:text-stone-600 transition-colors">
+          Advanced settings
+          <svg className="w-3 h-3" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9 5l7 7-7 7" />
+          </svg>
+        </button>
       </div>
     </div>
   );
